Rename component to remain consistent
