+++ packages/frontend/src/pages/common/AssetCell.tsx
