+++ packages/frontend/src/pages/positions/PositionDetails.tsx
@@ -1,7 +1,7 @@
 import { AssetId } from '@explorer/types'
 import React from 'react'
 
-import { AssetCell } from '../common/AssetNameCell'
+import { AssetCell } from '../common/AssetCell'
 import { Page } from '../common/Page'
 import { PageHeaderStats } from '../common/PageHeaderStats'
 import { SimpleLink } from '../common/SimpleLink'
